# api/openapi.yaml
info:
  title: PaySuper payment solution service.
  version: "1.0.0"
  description: Swagger Specification for PaySuper Payment Form API.
servers:
  - url: "https://checkout.pay.super.com"
    description: Production API
components:
  schemas:
...
    PaymentFormUserChangeLangRequest:
      type: object
      required:
      - order_id
      - lang
      properties:
        order_id:
          description: "The unique identifier for the order."
          type: string
        lang:
          description: "The language code."
          type: string
    PaymentFormUserChangePaymentAccountRequest:
      type: object
      required:
      - order_id
      - method_id
      - account
      properties:
        order_id:
          description: "The unique identifier for the order."
          type: string
        method_id:
          description: "The unique identifier for a payment method."
          type: string
        account:
          description: "The user's unique account in the merchant project."
          type: string
    ProcessBillingAddressRequest:
      type: object
      required:
      - order_id
      - country
      properties:
        order_id:
          description: "The unique identifier for the order."
          type: string
        country:
          description: "The user's country. Two-letter country code by ISO 3166-1, in uppercase (for instance\" U\")."
          type: string
        zip:
          description: "The user's ZIP code."
          type: string
...
    ResponseErrorMessage:
      type: object
      properties:
        code:
          description: "The error code."
          type: string
        message:
          description: "The error short description."
          type: string
        details:
          description: "The error details."
          type: string
    CreateOrderJsonProjectResponse:
      type: object
      properties:
        id:
          description: "The unique identifier for the order."
          type: string
        payment_form_url:
          description: "The URL of the PaySuper-hosted payment form."
          type: string
...
    ReCreateOrderRequest:
      type: object
      properties:
        order_id:
          description: "The unique identifier for the order."
          type: string
...
paths:
...
  /api/v1/order:
    post:
      summary: Create a payment order
      operationId: orderPathСreateJson
      description: Create a payment order with a user and order data
      tags:
      - Payment Order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderJsonProjectResponse"
        "400":
          description: The error code and message with the error details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/order/{order_id}:
    get:
      summary: Get an order data
      operationId: orderIdPathGetPaymentFormData
      description: Get an order data for rendering a payment form
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentFormJsonData"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/order/recreate:
    post:
      summary: Recreate a payment order
      operationId: orderReCreatePathRecreateOrder
      description: Recreate a payment order using the order UUID for the old order
      tags:
      - Order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReCreateOrderRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderJsonProjectResponse"
        "400":
          description: The error code and message with the error details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/{order_id}/language:
    patch:
      summary: Change a language
      operationId: orderLanguagePathChangeLanguage
      description: Change a language using the order ID
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentFormUserChangeLangRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentFormDataChangeResponseItem"
        "400":
          description: The error code and message with the error details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/{order_id}/customer:
    patch:
      summary: Change a customer
      operationId: orderCustomerPathChangeCustomer
      description: Change a customer using the order ID
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentFormUserChangePaymentAccountRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentFormDataChangeResponseItem"
        "400":
          description: The error code and message with the error details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/{order_id}/billing_address:
    post:
      summary: Change a billing address for the order
      operationId: orderBillingAddressPathProcessBillingAddress
      description: "Change a billing address for the order by the order's unique identifier"
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessBillingAddressRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessBillingAddressResponseItem"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/{order_id}/notify_sale:
    post:
      summary: Subscribe on sales or discounts notifications
      operationId: orderNotifySalesPathNotifySale
      description: Subscribe on sales or discounts notifications using the order ID
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserNotifyRequest"
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/{order_id}/notify_new_region:
    post:
      summary: Subscribe on notifications about new regions
      operationId: orderNotifyNewRegionPathNotifyNewRegion
      description: Subscribe on PaySuper newsletters about an initiation of a payment receiving from new regions
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserNotifyRequest"
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/{order_id}/platform:
    post:
      summary: Change an order platform
      operationId: orderPlatformPathChangePlatform
      description: Change an order platform by the order ID
      tags:
      - Order
      parameters:
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentFormUserChangePlatformRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentFormDataChangeResponseItem"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/orders/receipt/{receipt_id}/{order_id}:
    get:
      summary: Getting a payment receipt
      operationId: orderReceiptPathGetReceipt
      description: Getting a payment receipt data for rendering
      tags:
      - Order
      parameters:
      - name: receipt_id
        description: The unique identifier for the receipt
        in: path
        required: true
        schema:
          type: string
      - name: order_id
        description: The unique identifier for the order
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderReceipt"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/paylink/{id}:
    get:
      summary: Create a payment link
      operationId: paylinkIdPathGetOrderForPaylink
      description: Create a payment link by a paylink ID
      tags:
      - Order
      parameters:
      - name: id
        description: The unique identifier for the paylink
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
...
openapi: "3.0.2"